+++ packages/eyes-webdriverio-5/.mocharc.cvg.js
@@ -1,5 +1,6 @@
 const tags = {
   wd: [
+    'headfull',
     'webdriver',
     'mobile',
     'native',
@@ -16,12 +17,12 @@ const protocol = process.env.APPLITOOLS_WDIO_PROTOCOL in tags ? process.env.APPL
 module.exports = {
   spec: [
     './test/generic/*.spec.js',
-    '../sdk-shared/coverage-tests/custom/**/*.spec.js',
+    'node_modules/@applitools/sdk-shared/coverage-tests/custom/**/*.spec.js',
   ],
   parallel: true,
   jobs: 15,
   timeout: 0,
   reporter: 'spec-xunit-file',
-  require: ['../sdk-shared/coverage-tests/util/mocha-hooks.js'],
+  require: ['node_modules/@applitools/sdk-shared/coverage-tests/util/mocha-hooks.js'],
   grep: new RegExp(`^[\\w\\s]*?(\\((?:@(${tags[protocol].join('|')}) ?)+\\))?$`),
 }