+++ packages/eyes-storybook/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 
 
+## 3.12.1 - 2020/11/26
+
+- updated to @applitools/eyes-sdk-core@12.6.2 (from 12.6.0)
+- updated to @applitools/visual-grid-client@15.2.1 (from 15.2.0)
+
 ## 3.12.0 - 2020/11/19
 
 - updated to @applitools/dom-snapshot@4.2.7 (from 4.2.6)