+++ packages/eyes-webdriverio-4/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 
+## 2.25.0 - 2020/11/25
+
+- updated to @applitools/visual-grid-client@15.2.0 (from 15.1.1)
+
 ## 2.24.0 - 2020/11/10
 
 - deprecate `saveDebugData`