+++ packages/eyes-webdriverio-5-service/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 
+## 1.8.0 - 2020/11/26
+
+- updated to @applitools/eyes-webdriverio@5.27.0 (from 5.26.0)
+
 ## 1.7.0 - 2020/11/10
 
 - updated to @applitools/eyes-webdriverio@5.25.1 (from 5.24.0)