+++ packages/eyes-webdriverio-5/CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Unreleased
 
 
+## 5.27.0 - 2020/11/25
+
+- updated to @applitools/eyes-sdk-core@12.6.0 (from 12.5.7)
+- updated to @applitools/visual-grid-client@15.2.0 (from 15.1.1)
+
 ## 5.26.0 - 2020/11/10
 
 -- fix firefox region compensation issue