+++ packages/sdk-fake-eyes-server/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 
+## 2.1.1 - 2020/11/24
+
+- add `/resources/query/resources-exist` endpoint
+
 ## 2.1.0 - 2020/11/19
 
 - add `/job-info` endpoint