+++ packages/sdk-shared/CHANGELOG.md
@@ -3,6 +3,26 @@
 ## Unreleased
 
 
+## 0.5.16 - 2020/11/25
+
+- skipping tests
+
+## 0.5.15 - 2020/11/25
+
+- remove browser version from tags in generic tests
+
+## 0.5.14 - 2020/11/25
+
+- skipping headfull tests in CI
+
+## 0.5.13 - 2020/11/23
+
+- fixed check frame logic
+
+## 0.5.12 - 2020/11/22
+
+- github actions release
+
 ## 0.5.6 - 2020/11/19
 
 - updated to @applitools/sdk-coverage-tests@2.1.0 (from 2.0.4)